backend/app/requirements.txt
fastapi
uvicorn[standard]
PyPDF2
sentence-transformers
faiss-cpu
google-generativeai
python-multipart
pydantic
